backend/requirements2.txt
farm-haystack==1.26.4
transformers>=4.46,<5.0
langchain-core>=0.3.15,<0.4.0
langchain-huggingface==0.1.2
sentence-transformers>=2.6.0
tokenizers>=0.19.1
pydantic<2
nltk>3.8.1
protobuf>=3.20,<4.0
typeguard<3.0.0
torch==2.5.1
torchvision==0.20.1
tensorflow-intel==2.15.0
tensorflow-addons==0.22.0
streamlit==1.28.2
tensorboard==2.15.1
